More Art Assets, Sound Settings, Sound Effects

+++ Laplace/Assets/Scripts/Util/Singleton/AudioManager.cs
@@ -7,7 +7,7 @@ using UnityEngine;
 public class AudioManager : Singleton<AudioManager>
 {
     public AudioSource audioSource;
-    public int sfxVolume = 1;
+    public float sfxVolume = 1;
     public AudioClip bgm;
     void Awake()
     {
